+++ app/controllers/api/v1/omses/events_controller.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class Api::V1::OMSes::EventsController < Api::V1::ApplicationController
+  include DateHelper
   before_action :find_and_authorize_oms
   before_action :validate_from_timestamp, only: :index
   before_action :validate_limit, only: :index
@@ -19,7 +20,7 @@ class Api::V1::OMSes::EventsController < Api::V1::ApplicationController
 
   def validate_from_timestamp
     params.require(:from_timestamp)
-    @from_timestamp = params[:from_timestamp].to_datetime
+    @from_timestamp = from_timestamp(params[:from_timestamp])
   rescue ActionController::ParameterMissing, ArgumentError => e
     render json: { error: e }, status: 400
   end